+++ cmd/src/campaigns_preview.go
@@ -41,11 +41,7 @@ Examples:
 
 		_, url, err := campaignsExecute(ctx, out, svc, flags)
 		if err != nil {
-			out.Write("")
-			block := out.Block(output.Line("❌", output.StyleWarning, "Error"))
-			block.Write(err.Error())
-			block.Close()
-			out.Write("")
+			printExecutionError(out, err)
 			return &exitCodeError{nil, 1}
 		}